+++ modules/.submodules.json
@@ -1,7 +1,7 @@
 {
   "userId": [
     "digiTrustIdSystem",
-    "userId/id5IdSystem",
+    "id5IdSystem",
     "criteortusIdSystem"
   ],
   "adpod": [
